# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hex Auth API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /licenses/generate:
    post:
      operationId: generateLicense
      tags: [licenses]
      summary: Generate a new license key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseInput"
      responses:
        "201":
          description: License generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
  /licenses/redeem:
    post:
      operationId: redeemLicense
      tags: [licenses]
      summary: Redeem a license key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseRedeemInput"
      responses:
        "200":
          description: License redeemed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
        "400":
          description: Invalid or already used key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /licenses/{id}:
    delete:
      operationId: deleteLicense
      tags: [licenses]
      summary: Delete a license key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: License deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Logs
...
  /sdk/login:
    post:
      operationId: sdkLogin
      tags: [sdk]
      summary: SDK user authentication with HWID binding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SdkLoginInput"
      responses:
        "200":
          description: Authentication result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SdkAuthResult"
  /sdk/validate:
    post:
      operationId: sdkValidate
      tags: [sdk]
      summary: Validate an active session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SdkValidateInput"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SdkAuthResult"
  /sdk/variable/{name}:
    get:
      operationId: sdkGetVariable
      tags: [sdk]
      summary: Get a runtime variable value
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Variable value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SdkVariable"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        message:
          type: ["string", "null"]

    # Auth schemas
...
    License:
      type: object
      required: [id, key, status, plan, createdAt]
      properties:
        id:
          type: integer
        key:
          type: string
        status:
          type: string
        plan:
          type: string
        expiresAt:
          type: ["string", "null"]
        usedBy:
          type: ["integer", "null"]
        usedByUsername:
          type: ["string", "null"]
        createdAt:
          type: string
    LicenseInput:
      type: object
      required: [plan]
      properties:
        plan:
          type: string
        expiresAt:
          type: ["string", "null"]
        quantity:
          type: ["integer", "null"]
    LicenseRedeemInput:
      type: object
      required: [key]
      properties:
        key:
          type: string

    # Event log schemas
...
    SdkLoginInput:
      type: object
      required: [appId, username, password, hwid, version]
      properties:
        appId:
          type: string
        username:
          type: string
        password:
          type: string
        hwid:
          type: string
        version:
          type: string
    SdkValidateInput:
      type: object
      required: [sessionToken]
      properties:
        sessionToken:
          type: string
    SdkAuthResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        message:
          type: ["string", "null"]
        sessionToken:
          type: ["string", "null"]
        user:
          type: ["object", "null"]
          properties:
            username:
              type: string
            plan:
              type: string
            expiresAt:
              type: ["string", "null"]
    SdkVariable:
      type: object
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: string